collect common routines in new files, delete old files
