+++ seaice_implicit_solver.py
@@ -1,21 +1,19 @@
 import numpy as np
-# from numpy import linalg as npla
+
 from scipy.sparse import linalg as spla
 from scipy.sparse.linalg import LinearOperator
-# from scipy.optimize import fsolve
 
 from seaice_params import *
 from seaice_size import *
 
-from seaice_strainrates import strainrates
-from seaice_viscosities import viscosities
-from seaice_ocean_drag_coeffs import ocean_drag_coeffs
-from seaice_bottomdrag_coeffs import bottomdrag_coeffs
+from dynamics_routines import strainrates, viscosities, \
+    ocean_drag_coeffs, bottomdrag_coeffs, \
+    calc_stressdiv, calc_stress, calc_lhs, calc_rhs, calc_residual
+
 from seaice_global_sum import global_sum
 from seaice_fill_overlap import fill_overlap_uv
 from seaice_lsr import lsr_solver
 
-secondOrderBC = False
 nLinear = 50
 nonLinTol = 1e-5
 
@@ -28,127 +26,6 @@ nJFNK = 10
 computeJFNKResidual = True
 printJFNKResidual = True
 plotJFNKResidual = True
-
-
-def calc_stress(e11, e22, e12, zeta, eta, press, iStep, myTime, myIter):
-    from seaice_averaging import c_point_to_z_point
-    sig11 = zeta*(e11 + e22) + eta*(e11 - e22) - 0.5 * press
-    sig22 = zeta*(e11 + e22) - eta*(e11 - e22) - 0.5 * press
-    sig12 = 2. * e12 * c_point_to_z_point(eta)
-    return sig11, sig22, sig12
-
-def calc_stressdiv(sig11, sig22, sig12, iStep, myTime, myIter):
-    stressDivX = (
-          sig11*dyF - np.roll(sig11*dyF, 1,axis=1)
-        - sig12*dxV + np.roll(sig12*dxV,-1,axis=0)
-    ) * recip_rAw
-    stressDivY = (
-          sig22*dxF - np.roll(sig22*dxF, 1,axis=0)
-        - sig12*dyU + np.roll(sig12*dyU,-1,axis=1)
-    ) * recip_rAs
-    return stressDivX, stressDivY
-
-def calc_lhs(uIce, vIce, zeta, eta, press,
-             hIceMean, Area, areaW, areaS,
-             SeaIceMassC, SeaIceMassU, SeaIceMassV,
-             cDrag, cBotC, R_low,
-             iStep, myTime, myIter):
-
-    recip_deltaT = 1./deltaTdyn
-    bdfAlpha = 1.
-    sinWat = np.sin(np.deg2rad(waterTurnAngle))
-    cosWat = np.cos(np.deg2rad(waterTurnAngle))
-
-    #
-    e11, e22, e12          = strainrates(uIce, vIce, secondOrderBC)
-    # zeta, eta, press       = viscosities(
-    #     e11, e22, e12, press0, iStep, myIter, myTime)
-    sig11, sig22, sig12    = calc_stress(
-        e11, e22, e12, zeta, eta, press, iStep, myIter, myTime)
-    stressDivX, stressDivY = calc_stressdiv(
-        sig11, sig22, sig12, iStep, myIter, myTime)
-
-    # sum up for symmetric drag contributions
-    symDrag = cDrag*cosWat + cBotC
-
-    # mass*(uIce)/deltaT - dsigma/dx
-    uIceLHS = bdfAlpha*SeaIceMassU*recip_deltaT*uIce - stressDivX
-    # mass*(vIce)/deltaT - dsigma/dy
-    vIceLHS = bdfAlpha*SeaIceMassV*recip_deltaT*vIce - stressDivY
-    # coriols terms: - mass*f*vIce
-    #                + mass*f*uIce
-    fuAtC = SeaIceMassC * fCori * 0.5 * ( uIce + np.roll(uIce,-1,1) )
-    fvAtC = SeaIceMassC * fCori * 0.5 * ( vIce + np.roll(vIce,-1,0) )
-    uIceLHS = uIceLHS - 0.5 * ( fvAtC + np.roll(fvAtC,1,1) )
-    vIceLHS = vIceLHS + 0.5 * ( fuAtC + np.roll(fuAtC,1,0) )
-    # ocean-ice and bottom drag terms: + (Cdrag*cosWat+Cb)*uIce - vIce*sinWat)
-    uAtC = 0.5 * ( uIce + np.roll(uIce,-1,1) )
-    vAtC = 0.5 * ( vIce + np.roll(vIce,-1,0) )
-    uIceLHS = uIceLHS + (
-        0.5 * ( symDrag + np.roll(symDrag,1,1) ) *  uIce
-        - np.sign(fCori) * sinWat * 0.5 * (
-            symDrag * vAtC + np.roll(symDrag * vAtC,1,1)
-        )
-    ) * areaW
-    vIceLHS = vIceLHS + (
-        0.5 * ( symDrag + np.roll(symDrag,1,0) ) * vIce
-        + np.sign(fCori) * sinWat * 0.5 * (
-            symDrag * uAtC  + np.roll(symDrag * uAtC,1,0)
-        )
-    ) * areaS
-    # momentum advection (not now)
-
-    # apply masks for interior (important when we have open boundaries)
-    return uIceLHS*maskInW, vIceLHS*maskInS
-
-def calc_rhs(uIceRHSfix, vIceRHSfix, areaW, areaS,
-             uVel, vVel, cDrag,
-             iStep, myTime, myIter):
-
-    sinWat = np.sin(np.deg2rad(waterTurnAngle))
-    cosWat = np.cos(np.deg2rad(waterTurnAngle))
-
-    # ice-velocity independent contribution to drag terms
-    # - Cdrag*(uVel*cosWat - vVel*sinWat)/(vVel*cosWat + uVel*sinWat)
-    # ( remember to average to correct velocity points )
-    uAtC = 0.5 * ( uVel + np.roll(uVel,-1,1) )
-    vAtC = 0.5 * ( vVel + np.roll(vVel,-1,0) )
-    uIceRHS = uIceRHSfix + (
-        0.5 * ( cDrag + np.roll(cDrag,1,1) ) * cosWat *  uVel
-        - np.sign(fCori) * sinWat * 0.5 * (
-            cDrag * vAtC + np.roll(cDrag * vAtC,1,1)
-        )
-    ) * areaW
-    vIceRHS = vIceRHSfix + (
-        0.5 * ( cDrag + np.roll(cDrag,1,0) ) * cosWat * vVel
-        + np.sign(fCori) * sinWat * 0.5 * (
-            cDrag * uAtC  + np.roll(cDrag * uAtC,1,0)
-        )
-    ) * areaS
-
-    return uIceRHS*maskInW, vIceRHS*maskInS
-
-def calc_residual(uIce, vIce, hIceMean, Area,
-                  zeta, eta, press, cDrag, cBotC,
-                  SeaIceMassC, SeaIceMassU, SeaIceMassV,
-                  forcingU, forcingV, uVel, vVel, R_low,
-                  iStep, myTime, myIter):
-
-    # initialize fractional areas at velocity points
-    areaW = 0.5 * (Area + np.roll(Area,1,1))
-    areaS = 0.5 * (Area + np.roll(Area,1,0))
-
-    Au, Av = calc_lhs(uIce, vIce, zeta, eta, press,
-                      hIceMean, Area, areaW, areaS,
-                      SeaIceMassC, SeaIceMassU, SeaIceMassV,
-                      cDrag, cBotC, R_low,
-                      iStep, myIter, myTime)
-    bu, bv = calc_rhs(forcingU, forcingV, areaW, areaS,
-                      uVel, vVel, cDrag,
-                      iStep, myIter, myTime)
-
-    # residual by vector component
-    return Au-bu, Av-bv
 
 def _vecTo2d(vec):
     n = vec.shape[0]//2
@@ -301,7 +178,7 @@ def picard_solver(uIce, vIce, uVel, vVel, hIceMean, Area,
         #
         cDrag = ocean_drag_coeffs(uIceLin, vIceLin, uVel, vVel)
         cBotC = bottomdrag_coeffs(uIceLin, vIceLin, hIceMean, Area, R_low)
-        e11, e22, e12    = strainrates(uIceLin, vIceLin, secondOrderBC)
+        e11, e22, e12    = strainrates(uIceLin, vIceLin)
         zeta, eta, press = viscosities(
             e11, e22, e12, press0, iPicard, myIter, myTime)
         bu, bv = calc_rhs(uIceRHSfix, vIceRHSfix, areaW, areaS,
@@ -400,18 +277,19 @@ def jfnk_solver(uIce, vIce, uVel, vVel, hIceMean, Area,
     areaS = 0.5 * (Area + np.roll(Area,1,0))
     exitCode = 1
     iJFNK = -1
-    resNonLin = nonLinTol*2
+    resNonLin = 1.
     linTol = 1e-1
-    while resNonLin > nonLinTol and iJFNK < nJFNK:
+    JFNKtol = 0.
+    while resNonLin > JFNKtol and iJFNK < nJFNK:
         iJFNK = iJFNK+1
         # smoothing
-        wght=.8
+        wght=1
         uIceLin = wght*uIce+(1.-wght)*uIceLin
         vIceLin = wght*vIce+(1.-wght)*vIceLin
         #
         cDrag = ocean_drag_coeffs(uIceLin, vIceLin, uVel, vVel)
         cBotC = bottomdrag_coeffs(uIceLin, vIceLin, hIceMean, Area, R_low)
-        e11, e22, e12    = strainrates(uIceLin, vIceLin, secondOrderBC)
+        e11, e22, e12    = strainrates(uIceLin, vIceLin)
         zeta, eta, press = viscosities(
             e11, e22, e12, press0, iJFNK, myIter, myTime)
         Fu, Fv = calc_residual(uIce, vIce, hIceMean, Area,
@@ -446,6 +324,9 @@ def jfnk_solver(uIce, vIce, uVel, vVel, hIceMean, Area,
         M = preconGmres(u, J)
 
         resNonLin = np.sqrt( (Fu**2+Fv**2).sum() )
+        if iJFNK==0:
+            JFNKtol = resNonLin*nonLinTol
+            print('JFNKtol = %e'%JFNKtol)
 
         if printJFNKResidual:
             print(
